fix: moved functions to their "right" place
